Fix auth denying access on simultaneous requests
useEffect with React Strict mode exposed an issue in the auth
middleware where multiple requests sent in very close proximity would
cause a 401 response from the server.

Previously, the middleware
immediately expired the refresh token it received as a cookie with the
request, then generated a new one and sent that as a response. However
if multiple requests were sent roughly at the same time from the client,
they would all include the "old" refresh token, causing the server to
reply with a 401 to all but the first one.

This change will sets the expiry date of a received refresh token to
10 seconds, instead of expiring it immediately. That gives the network
and server enough time to process that batch of requests.

Also implemented functionality for removing expired tokens from the DB
to prevent the array from growing to excessive proportions.