Add TempUser model to mongoose
This model stores a temporary user object created after the initial
signup request and before they clicked the link inside the verification
email. The Mongo Atlas deployment has a TTL index on the expDate field
that makes sure TempUser docs are removed after 24 hours, regardless if
the user confirmed their email or not.