Add 'userId' to auth JWT payload
In the inital concept of the API, a user's email was supposed to be
their primary unique identifier. However, mongoose's "populate"
functionality requires the use of the _id field instead. The "populate"
method is handy enough that I ended up relying on _id's instead of
emails for identifying users in most places across the app, including
exposing them on the client-side.
Since most controller function rely on the _id field, the auth JWT lost
its main advantage - the reduction in the amount of DB fetches for the
user object. This commit changes that by adding the userId field into
the auth JWT, allowing functions to directly pull the _id from the JWT
instead of fetching the user obj using "email" first.