Add req.context object
Added simple middleware that sets the "context" object on incoming
requests and typescript definitions for some of the data that can be
stored inside the context object.

The purpose of this change is to allow sharing of data between
middleware and controllers. This can help improve performance e.g.
by allowing the sharing of the "requestingUser" object which gets rid
of an unnecessary fetch for User.