Fix patchDoc getting stuck when patching server
The "flatten" function from 'mongo-dot-notation' was getting stuck
while processing a subset of a Server document. The issue was caused
by some of the virtual properties inside nested mongoose documents.

getPatchableServer was modified to use the .toObject() method which
returns a plain object, without all the virtual properties. This fixes
the issue.