Refactor updateServer servers controller
Since validation middleware has no way of knowing if the patch document
includes the "serverImg" property, it can't be responsible for making
sure that the serverImg file is present in the request. This change
adds that check into the controller function, so that the update
request doesn't get processed if the serverImg file is missing.