Refactor removeImage images service function
removeImage function was previously accepting "pathname" to the image
as an argument, which meant that all the users of the function had to
know how to extract the "pathname" of the image object.
The new implementation makes use of the universal ImageObject interface
which allows it to extract the pathname itself, freeing the consumers
of the function from that responsibility and thus decoupling the code.