Rename SERVER_INVITE_PATH client config constant
Renamed SERVER_INVITE_PATH to CLIENT_SERVER_INVITE_PATH. This is meant
to achieve consistency with other constants exported by client config.
The reason behind this seemingly redundant naming convention is that
when imported inside code, it's immediately obvious if the e.g.
"SERVER_INVITE_PATH" refers to the path on the client side or in the
api.