Fix getMessages controller returned value
getMessages chat controller function was returning chat messages with
a populated "author" field that included "profileImg" as an ImageObject
instead of a URL to the image asset. The controller has been changed
to run the message docs through the getClientSafeSubset chat service
function, which resolves all the getters, including the one for
"profileImg", and fixes the problem.