Fix side effect in getUserServersIn
In order to obtain the populated serversIn array, the function had to
call populate() method on the received User doc. However the populate
method works in-place, meaning the doc would be mutated without
notifying the user, not correctly updating the document's types. This
change depopulates the field after the serversIn array has been
obtained.