Fix serversIn populate not persisting
The "populateServersIn" option in getUserServersIn users service
method wasn't properly populating the array. The issue was called
by a depopulate call which modifies the array in-place, which MessageCollection
it also affected the array returned by the method.

To fix the problem, the returned array has been turned into a shallow
copy of the original array.