Turn setInfisicalToken into entrypoint script
setInfisicalToken didn't work properly because the environment variable
it set was limited to the scope of the script and wasn't available
to Infisical tool. The script was reworked and renamed into general
entrypoint script, which uses "exec" command to persist its environment
variables and make them available inside the "npm run start" script.