Fix demo users duplicate key error
When dummy accounts of fake demo users are created and pushed into the
db, the value of emailVerificationTokens property is an empty array,
since they don't need to verify their email.
EmailVerificationToken objects which are supposed to populate that
array, have a unique index defined as part of their schema. This led to
an issue, since Mongo's default behavior of creating a unique index
on empty arrays is to set the indexed field to "null" value. With
multiple dummy demo accounts, this led to a situation where multiple
documents were being pushed to DB with the same "null" value, causing
the unique index to interpret them as duplicates and throwing the
E11000 mongo error.

The schema has been reconfigured to use a sparse unique index instead,
which ignores null values and fixes the problem.