Fix nodemailer templates missing in production
Typescript compiler wasn't including .hbs email template files in the
compiled build, which caused issues in production. Also nodemailer
was configured with hard-coded paths to mail templates in the src
directory.

Mail templates have been moved out of the src directory and into the
root directory which solves both issues, since the templates are
available at the same path regardless of the environment the app is
running in.